Refactor shared components
