+++ web/src/app/auth/auth.module.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 import { StoreModule } from '@ngrx/store';
 import { WebAuth } from 'auth0-js';
-
-import { StorageModule } from '../storage';
-
 import { environment } from '../../environments/environment';
+import { StorageModule } from '../storage';
 
 @NgModule({
   imports: [
