+++ web/src/app/content/content.module.ts
@@ -2,17 +2,14 @@ import { CommonModule } from '@angular/common';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-
-import { ContentComponent } from './content.component';
-import { ContentRoutingModule } from './content-routing.module';
-
-import { ArchiveComponent } from './archive';
-import { RecentPostsComponent } from './recent-posts';
-
-import { AppHttpInterceptor } from '../index';
 import { AuthModule } from '../auth';
 import { GraphQLModule } from '../graphql';
+import { AppHttpInterceptor } from '../index';
 import { SharedModule } from '../shared';
+import { ArchiveComponent } from './archive';
+import { ContentRoutingModule } from './content-routing.module';
+import { ContentComponent } from './content.component';
+import { RecentPostsComponent } from './recent-posts';
 
 @NgModule({
   imports: [
