+++ web/src/app/content/index.ts
@@ -1,4 +1,5 @@
-export * from './content.component';
 export * from './content-routing.module';
+export * from './content.component';
 export * from './content.module';
 export * from './recent-posts';
+
