+++ web/src/app/graphql/graphql.module.ts
@@ -1,12 +1,10 @@
-import { NgModule } from '@angular/core';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { NgModule } from '@angular/core';
 import { ApolloModule, APOLLO_OPTIONS } from 'apollo-angular';
-import { HttpLinkModule, HttpLink } from 'apollo-angular-link-http';
+import { HttpLink, HttpLinkModule } from 'apollo-angular-link-http';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
-
 import { environment } from 'src/environments/environment';
-
 import { AppHttpInterceptor } from '../app-http.interceptor';
 
 const uri = environment.graphql.endpoint;
