+++ web/src/app/shared/button/button.module.ts
@@ -1,10 +1,10 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { ButtonComponent } from './button';
+import { DropdownButtonComponent } from './dropdown';
+import { OutlineButtonComponent } from './outline';
 
-import { ButtonComponent } from './button.component';
-import { DropdownButtonComponent } from './dropdown-button.component';
-import { OutlineButtonComponent } from './outline-button.component';
 
 @NgModule({
   imports: [
