+++ web/src/app/shared/button/button/button.component.scss
@@ -1,0 +1,20 @@
+:host {
+  align-items: center;
+  cursor: pointer;
+  display: inline-flex;
+  height: 100%;
+  justify-content: center;
+}
+
+.button {
+  background: none;
+  border: none;
+  color: #808080;
+  font: normal 400 1.3rem Lato sans-serif;
+  padding: 0;
+}
+
+.icon {
+  font-size: 2.2rem;
+  padding: 0 3.2rem;
+}
