+++ web/src/app/shared/button/button/index.ts
@@ -1,0 +1,1 @@
+export * from './button.component';
