+++ web/src/app/shared/button/dropdown/dropdown.component.html
@@ -1,0 +1,15 @@
+<button class="button" [attr.aria-label]="ariaLabel" (click)="toggleMenu()">
+
+  <span [style.margin-left]="'auto'"></span>
+
+  <span>{{selectedItem.label}}</span>
+
+  <fa-icon class="icon" [icon]="faChevronDown"></fa-icon>
+
+</button>
+
+<ul class="dropdown-menu _list-unstyled" [@slideUpDown]="showMenu ? true : false">
+
+  <li class="menu-item" *ngFor="let item of items" (click)="onClick(item);toggleMenu()">{{item.label}}</li>
+
+</ul>
