+++ web/src/app/shared/button/dropdown/dropdown.component.scss
@@ -1,0 +1,59 @@
+:host {
+  display: block;
+  position: relative;
+}
+
+.button {
+  align-items: center;
+  background: #0288d1;
+  border: none;
+  border-radius: 100vh;
+  color: #fff;
+  display: flex;
+  font: normal 400 1.3rem Lato sans-serif;
+  height: 3.2rem;
+  padding: 0 1.6rem;
+  width: 100%;
+}
+
+.icon {
+  font-size: 0.8rem;
+  margin-left: auto;
+}
+
+.icon:after {
+  clear: both;
+}
+
+.dropdown-menu {
+  background: #fff;
+  border: 0.1rem solid #ececec;
+  border-radius: 0.4rem;
+  padding: 1.6rem 0 !important;
+  position: absolute;
+  right: 0;
+  top: 4.8rem;
+  width: 100%;
+  z-index: 98;
+}
+
+.dropdown-menu:before {
+  background: #fff;
+  border-top: 0.1rem solid #ececec;
+  border-left: 0.1rem solid #ececec;
+  content: "";
+  display: block;
+  height: 1.6rem;
+  position: absolute;
+  right: 0.8rem;
+  width: 1.6rem;
+  top: -0.8rem;
+  transform: rotate(45deg);
+}
+
+.menu-item {
+  color: #333;
+  cursor: pointer;
+  font: normal 400 1.6rem Lato sans-serif;
+  padding: 0.8rem 1.6rem;
+}
