+++ web/src/app/shared/button/dropdown/index.ts
@@ -1,0 +1,1 @@
+export * from './dropdown.component';
