+++ web/src/app/shared/button/index.ts
@@ -1,0 +1,4 @@
+export * from './button';
+export * from './button.module';
+export * from './dropdown';
+export * from './outline';
