+++ web/src/app/shared/button/outline/index.ts
@@ -1,0 +1,1 @@
+export * from './outline.component';
