+++ web/src/app/shared/button/outline/outline.component.html
@@ -1,0 +1,5 @@
+<button class="button" [attr.aria-label]="ariaLabel">
+
+  <ng-content></ng-content>
+
+</button>
