+++ web/src/app/shared/button/outline/outline.component.ts
@@ -1,0 +1,10 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ButtonComponent } from '../button';
+
+@Component({
+  selector: 'app-outline-button',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  templateUrl: './outline.component.html',
+  styleUrls: ['./outline.component.scss'],
+})
+export class OutlineButtonComponent extends ButtonComponent { }
