+++ web/src/app/shared/dialog/dialog.component.html
@@ -1,0 +1,1 @@
+<ng-content></ng-content>
