+++ web/src/app/shared/dialog/dialog.module.ts
@@ -1,11 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-
 import { DialogComponent } from './dialog.component';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
   ],
   declarations: [
     DialogComponent,
