+++ web/src/app/shared/dialog/index.ts
@@ -1,0 +1,2 @@
+export * from './dialog.component';
+export * from './dialog.module';
