+++ web/src/app/shared/share.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
-
-import { ButtonModule } from './button/button.module';
-import { DialogModule } from './dialog/dialog.module';
+import { ButtonModule } from './button';
+import { DialogModule } from './dialog';
 import { PostModule } from './post/post.module';
 import { TemplateModule } from './template/template.module';
 import { UserModule } from './user/user.module';
