+++ web/src/app/shared/user/index.ts
@@ -1,0 +1,2 @@
+export * from './navbar';
+export * from './user.module';
