+++ web/src/app/shared/user/navbar/index.ts
@@ -1,0 +1,1 @@
+export * from './navbar.component';
