+++ web/src/app/shared/user/navbar/navbar.component.scss
