--- web/src/app/shared/user/user.component.ts
@@ -1,8 +1,0 @@
-import { Input } from '@angular/core';
-
-export abstract class UserComponent {
-
-  @Input()
-  userInfo: UserInfo;
-
-}
